Convet m4s (Steam recodings) to mp4 via ffmpeg
The export of video files recorded via Steam works very poorly. This is likely because Steam doesn't utilize any hardware/GPU acceleration (and seems to encode in a single thread). On a new MBP machines, this becomes downright ridiculous — an 11 minute clip takes over 2 hours to compress.

As a result, this utility should significantly speed up the entire process.